fix: 共通widgetのフォルダに移動
