+++ lib/presentation/searchbook/searchbook_page.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:skeleton_text/skeleton_text.dart';
-import 'package:image_network/image_network.dart';
 import '../../presentation/searchbook/searchbook_page_notifier.dart';
-import '../../domain/entity/book.dart';
+import '../../presentation/widget/BookBox.dart';
 
 class SearchBookPage extends HookConsumerWidget {
   const SearchBookPage({Key? key}) : super(key: key);
@@ -45,7 +43,7 @@ class SearchBookPage extends HookConsumerWidget {
                 itemCount: state.books.length,
                 itemBuilder: (context, index) {
                   final item = state.books[index];
-                  return _BookBox(
+                  return BookBox(
                     book: item,
                   );
                 },
@@ -54,83 +52,6 @@ class SearchBookPage extends HookConsumerWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-class _BookBox extends StatelessWidget {
-  const _BookBox({required this.book, Key? key}) : super(key: key);
-  final Book book;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        Row(
-          children: [
-            Expanded(
-              flex: 2,
-              child: Container(
-                margin: const EdgeInsets.only(left: 10, right: 5),
-                child: ImageNetwork(
-                  image: book.smallImageUrl ?? '',
-                  height: 60,
-                  width: 50,
-                  fitAndroidIos: BoxFit.contain,
-                  onLoading: SkeletonAnimation(
-                    child: Container(
-                      color: const Color.fromARGB(70, 125, 125, 125),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-            Expanded(
-              flex: 9,
-              child: Column(
-                children: [
-                  Container(
-                    alignment: Alignment.topLeft,
-                    child: Text(
-                      book.title ?? '',
-                      style: const TextStyle(
-                        color: Color.fromARGB(255, 0, 0, 0),
-                        fontSize: 16,
-                      ),
-                    ),
-                  ),
-                  Container(
-                    alignment: Alignment.topLeft,
-                    child: Text(
-                      book.author ?? '',
-                      style: const TextStyle(
-                        color: Color.fromARGB(255, 184, 180, 180),
-                        fontSize: 16,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            const Expanded(
-              flex: 1,
-              child: Icon(
-                Icons.add_box,
-                color: Colors.black38,
-                size: 30,
-              ),
-            ),
-          ],
-        ),
-        const Divider(
-          color: Colors.black12,
-          thickness: 3,
-          height: 10,
-          indent: 10,
-        ),
-      ],
     );
   }
 }
